+++ app/components/Education.tsx
@@ -4,8 +4,8 @@ export const Education = (_props: EducationProps) => {
   return (
     <section>
       <div>
-        <h2 className="py-8 text-lg md:text-xl">Education</h2>
-        <div className="mb-8 border-t border-solid border-color-border" />
+        <h2 className="py-8 text-lg print:py-4 md:text-xl">Education</h2>
+        <div className="mb-8 border-t border-solid border-color-border print:hidden" />
       </div>
 
       <h3>San Francisco State University</h3>
