+++ app/components/Experience.tsx
@@ -1,82 +1,41 @@
-import type { Experience as ExperienceType } from "~/data/resume";
-import { getTimeWorked } from "~/utils/date-time";
+import { useState } from "react";
+import { experience as data } from "~/data/resume";
+import { ExperienceDetail } from "./ExperienceDetail";
 
-export interface ExperienceProps {
-  experience: ExperienceType;
-}
-
-export const Experience = (props: ExperienceProps) => {
-  const { experience } = props;
-  const {
-    company,
-    companyUrl,
-    date,
-    dateRange,
-    description,
-    highlights,
-    image,
-    title
-  } = experience;
+export const Experience = () => {
+  // Hooks
+  // const minValue = data.length;
+  const minValue = 3;
+  const [shown, setShown] = useState(minValue);
 
   // Setup
-  const start = dateRange ? dateRange[0] : new Date();
-  const stop = dateRange?.[1] ? dateRange[1] : new Date();
-  const length = getTimeWorked(start, stop);
+  const experience = data.slice(0, shown);
 
-  // TEMP
-  console.log(` 💬 ~ Worked at ${company} for`, length);
+  // Handlers
+  const onToggleExp = () => {
+    setShown(shown === minValue ? data.length : minValue);
+  };
 
   return (
-    <div className="flex flex-col gap-10 md:flex-row" key={title}>
-      <div className="flex flex-col gap-2 text-base md:basis-[160px]">
-        <div className="flex">
-          <h3 className="flex items-center gap-2 font-bold">
-            <a
-              className="underline-offset-4 hover:underline print:text-color-copy-dark"
-              dangerouslySetInnerHTML={{ __html: company }}
-              href={companyUrl}
-              rel="noreferrer"
-              target="_blank"
-            />
-            {experience.contract && (
-              <small className="text-xs font-normal italic text-gray-700">
-                (contract)
-              </small>
-            )}
-            {image && (
-              <img
-                alt={`${company} favicon`}
-                className="h-4 w-4"
-                height="auto"
-                loading="eager"
-                src={image}
-                width="auto"
-              />
-            )}
-          </h3>
-        </div>
-        <h3 className="uppercase- text-xs font-medium">
-          <div className="uppercase">{title}</div>
-          <div className="text-color-copy-light">{date}</div>
-        </h3>
+    <section>
+      <div>
+        <h2 className="py-8 text-lg print:py-4 md:text-xl">Experience</h2>
+        <div className="mb-8 border-t border-solid border-color-border print:hidden" />
       </div>
 
-      <div className="flex-1 text-sm leading-4">
-        <div
-          dangerouslySetInnerHTML={{
-            __html: description
-          }}
-        />
-        <ul className="my-4 ml-4 list-disc text-sm font-light text-color-copy">
-          {highlights.map((highlight, index) => (
-            <li
-              className="my-1"
-              key={index}
-              dangerouslySetInnerHTML={{ __html: highlight }}
-            />
-          ))}
-        </ul>
+      <div className="mb-10 flex flex-col gap-10">
+        {experience.map((exp) => (
+          <ExperienceDetail experience={exp} key={exp.title} />
+        ))}
       </div>
-    </div>
+      <div className="print:hidden">
+        <button
+          className="m-10 mx-auto block rounded-2xl border border-color-border bg-color-background-light py-2 px-4 font-font-sans-serif text-xs font-bold uppercase"
+          onClick={onToggleExp}
+        >
+          {shown === minValue ? "Show more" : "Show less"}
+        </button>
+      </div>
+    </section>
   );
 };
