+++ app/components/Profeciencies.tsx
@@ -4,8 +4,8 @@ export const Proficiencies = () => {
   return (
     <section>
       <div>
-        <h2 className="py-8 text-lg md:text-xl">Proficiencies</h2>
-        <div className="mb-8 border-t border-solid border-color-border" />
+        <h2 className="py-8 text-lg print:py-4 md:text-xl">Proficiencies</h2>
+        <div className="mb-8 border-t border-solid border-color-border print:hidden" />
       </div>
 
       <p className="text-sm">
@@ -14,43 +14,43 @@ export const Proficiencies = () => {
         to life.
       </p>
 
-      <div className="mt-8 grid grid-cols-1 gap-8 p-1 text-sm md:grid-cols-2">
-        <div className="rounded-md border bg-color-background-light p-4 shadow-md">
+      <div className="mt-8 grid grid-cols-1 gap-8 p-1 text-sm print:grid-cols-3 md:grid-cols-2">
+        <div className="rounded-md border bg-color-background-light p-4 shadow-md print:border-0 print:shadow-none">
           <h3 className="text-lg text-color-copy-light">DevOps</h3>
         </div>
-        <div className="rounded-md border bg-color-background-light p-4 shadow-md">
+        <div className="rounded-md border bg-color-background-light p-4 shadow-md print:border-0 print:shadow-none">
           <h3 className="text-lg text-color-copy-light">Testing</h3>
           Redux, REST, Rollup, SQL / MySQL, Sketch, Swarm, Tailwind CSS,
           TypeScript, Webpack,
         </div>
-        <div className="rounded-md border bg-color-background-light p-4 shadow-md">
+        <div className="rounded-md border bg-color-background-light p-4 shadow-md print:border-0 print:shadow-none">
           <h3 className="text-lg text-color-copy-light">Testing</h3>
           HTML, Javascript, JWT, Kubernetes, Lerna, NextJS, Node.js, NoSQL, npm,
           PostCSS, Postgres, React, Remix 🤩, RecoilJS,
         </div>
-        <div className="rounded-md border bg-color-background-light p-4 shadow-md">
+        <div className="rounded-md border bg-color-background-light p-4 shadow-md print:border-0 print:shadow-none">
           <h3 className="text-lg text-color-copy-light">Design</h3>
           Adobe; you name it I've likely used it, Sketch, Figma
         </div>
-        <div className="rounded-md border bg-color-background-light p-4 shadow-md">
+        <div className="rounded-md border bg-color-background-light p-4 shadow-md print:border-0 print:shadow-none">
           <h3 className="text-lg text-color-copy-light">Testing</h3>
           Apollo, Agile, AWS, Babel, CSS, Docker, ElasticSearch, Express, Figma,
           Google Analytics, Google Cloud, Google Tag Manger, Git/GitHub,
           GraphQL,
         </div>
-        <div className="rounded-md border bg-color-background-light p-4 shadow-md">
+        <div className="rounded-md border bg-color-background-light p-4 shadow-md print:border-0 print:shadow-none">
           <h3 className="text-lg text-color-copy-light">Testing</h3>
           Apollo, Agile, AWS, Babel, CSS, Docker, ElasticSearch, Express, Figma,
           Google Analytics, Google Cloud, Google Tag Manger, Git/GitHub,
           GraphQL,
         </div>
-        <div className="rounded-md border bg-color-background-light p-4 shadow-md">
+        <div className="rounded-md border bg-color-background-light p-4 shadow-md print:border-0 print:shadow-none">
           <h3 className="text-lg text-color-copy-light">Testing</h3>
           Apollo, Agile, AWS, Babel, CSS, Docker, ElasticSearch, Express, Figma,
           Google Analytics, Google Cloud, Google Tag Manger, Git/GitHub,
           GraphQL,
         </div>
-        <div className="rounded-md border bg-color-background-light p-4 shadow-md">
+        <div className="rounded-md border bg-color-background-light p-4 shadow-md print:border-0 print:shadow-none">
           <h3 className="text-lg text-color-copy-light">Testing</h3>
           Apollo, Agile, AWS, Babel, CSS, Docker, ElasticSearch, Express, Figma,
           Google Analytics, Google Cloud, Google Tag Manger, Git/GitHub,
